+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[Fix]** Fix Typedoc generation: use `tsconfig.json` generated for the lib.
 - **[Fix]** Write source map for `.mjs` files
 - **[Fix]** Copy sources to `_src` when publishing a lib (#87).
+- **[Internal]** Restore continuous deployment of documentation.
 
 ## 0.16.1 (2017-01-20)